Clean up storybook code
