+++ stories/custom_components/selection_status.js
@@ -1,0 +1,16 @@
+import React from "react";
+import Avatar from "material-ui/Avatar";
+
+const SelectionStatus = ({ selected }) => (
+  <div
+    style={{
+      display: "flex",
+      justifyContent: "flex-end",
+      padding: "0 12px"
+    }}
+  >
+    <Avatar>{selected.length}</Avatar>
+  </div>
+);
+
+export default SelectionStatus;
