+++ stories/multi-select.stories.js
@@ -1,13 +1,13 @@
-import React, { Component } from "react";
+import React from "react";
 import { action, storiesOf } from "@storybook/react";
 import ReactMultiSelect from "../src/components/multi_select";
 import { boolean, withKnobs } from "@storybook/addon-knobs";
-import TextField from "material-ui/TextField";
-import StarOIcon from "react-icons/lib/md/star-outline";
-import StarIcon from "react-icons/lib/md/star";
-import DeleteIcon from "react-icons/lib/md/delete";
-import Avatar from "material-ui/Avatar";
 import customStyle from "./custom_style.scss";
+import SelectAll from "./custom_components/select_all";
+import SelectionStatus from "./custom_components/selection_status";
+import Search from "./custom_components/search";
+import Item from "./custom_components/item";
+import SelectedItem from "./custom_components/selected_item";
 
 const custom_messages = {
   searchPlaceholder: "Find...",
@@ -18,14 +18,19 @@ const custom_messages = {
   clearAllMessage: "Uncheck all"
 };
 
+const generateItems = size =>
+  Array.apply(null, { length: size }).map((i, index) => ({
+    id: index,
+    label: `Item ${index}`
+  }));
+
+const items = generateItems(50);
+
+const manyItems = generateItems(7000);
+
 storiesOf("React Multi Select", module)
   .addDecorator(withKnobs)
   .add("Default view", () => {
-    const items = Array.apply(null, { length: 10 }).map((i, index) => ({
-      id: index,
-      label: `Item ${index}`
-    }));
-
     return (
       <ReactMultiSelect
         items={items}
@@ -37,11 +42,6 @@ storiesOf("React Multi Select", module)
     );
   })
   .add("Preselected Items", () => {
-    const items = Array.apply(null, { length: 10 }).map((i, index) => ({
-      id: index,
-      label: `Item ${index}`
-    }));
-
     return (
       <ReactMultiSelect
         items={items}
@@ -54,13 +54,6 @@ storiesOf("React Multi Select", module)
     );
   })
   .add("With max selected items", () => {
-    const items = Array.apply(null, { length: 10 }).map((i, index) => {
-      return {
-        id: index,
-        label: `Item ${index}`
-      };
-    });
-
     return (
       <ReactMultiSelect
         items={items}
@@ -73,13 +66,6 @@ storiesOf("React Multi Select", module)
     );
   })
   .add("With Custom Messages", () => {
-    const items = Array.apply(null, { length: 10 }).map((i, index) => {
-      return {
-        id: index,
-        label: `Item ${index}`
-      };
-    });
-
     return (
       <ReactMultiSelect
         items={items}
@@ -92,13 +78,6 @@ storiesOf("React Multi Select", module)
     );
   })
   .add("With Custom Styling", () => {
-    const items = Array.apply(null, { length: 50 }).map((i, index) => {
-      return {
-        id: index,
-        label: `Item ${index}`
-      };
-    });
-
     return (
       <ReactMultiSelect
         items={items}
@@ -113,13 +92,6 @@ storiesOf("React Multi Select", module)
     );
   })
   .add("Without Search and Select all", () => {
-    const items = Array.apply(null, { length: 50 }).map((i, index) => {
-      return {
-        id: index,
-        label: `Item ${index}`
-      };
-    });
-
     return (
       <ReactMultiSelect
         items={items}
@@ -133,16 +105,9 @@ storiesOf("React Multi Select", module)
     );
   })
   .add("With Large Data (7000 items)", () => {
-    const items = Array.apply(null, { length: 7000 }).map((i, index) => {
-      return {
-        id: index,
-        label: `Item ${index}`
-      };
-    });
-
     return (
       <ReactMultiSelect
-        items={items}
+        items={manyItems}
         loading={boolean("Loading", false)}
         onChange={action("onChange")}
         showSearch={boolean("Show search", true)}
@@ -151,84 +116,6 @@ storiesOf("React Multi Select", module)
     );
   })
   .add("With custom components", () => {
-    const items = Array.apply(null, { length: 10 }).map((i, index) => ({
-      id: index,
-      value: `item ${index}`,
-      label: `Item ${index}`
-    }));
-
-    const Item = ({ item, checked }) => (
-      <div
-        style={{
-          display: "flex",
-          alignItems: "center",
-          lineHeight: "14px",
-          height: "100%",
-          margin: "12px"
-        }}
-      >
-        <div style={{ fontSize: "24px", margin: "0 12px 0 0" }}>
-          {!checked ? <StarOIcon /> : <StarIcon />}
-        </div>
-        {item.label}
-      </div>
-    );
-
-    const Search = ({ searchPlaceholder, onChange }) => {
-      return (
-        <TextField
-          placeholder={searchPlaceholder}
-          onChange={onChange}
-          fullWidth
-          style={{ margin: "17px 0 0 0" }}
-        />
-      );
-    };
-
-    const SelectAll = ({ isAllSelected, onClick }) => (
-      <div
-        onClick={onClick}
-        style={{
-          height: "40px",
-          background: "#e6e4ff",
-          padding: "0 12px",
-          display: "flex",
-          alignItems: "center"
-        }}
-      >
-        {isAllSelected ? "Unselect All" : "Select All"}
-      </div>
-    );
-
-    const SelectedItem = ({ item, height }) => (
-      <div
-        style={{
-          height,
-          display: "flex",
-          alignItems: "center",
-          lineHeight: "14px",
-          padding: "0 12px"
-        }}
-      >
-        <div>{item.label}</div>
-        <div style={{ fontSize: "24px", margin: "0 12px" }}>
-          <DeleteIcon />
-        </div>
-      </div>
-    );
-
-    const SelectionStatus = ({ selected }) => (
-      <div
-        style={{
-          display: "flex",
-          justifyContent: "flex-end",
-          padding: "0 12px"
-        }}
-      >
-        <Avatar>{selected.length}</Avatar>
-      </div>
-    );
-
     return (
       <ReactMultiSelect
         items={items}
